Added support to 3DUI, sawOpenIGTLink, and ui3exDaVinciStereoVideo for inlaid svl for additional Slicer renderer (PiP), with tool tracking. To run you will need to set flag, VOLUME_RENDERING 1, in ui3exDaVinciStereoVideo example.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2012-06-06-quadro-sdi@4611 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisst3DUserInterface/code/ui3BehaviorBase.cpp
@@ -210,15 +210,15 @@ void ui3BehaviorBase::SetState(const StateType & newState)
 }
 
 
-void ui3BehaviorBase::OnStreamSample(svlSample * CMN_UNUSED(sample), size_t CMN_UNUSED(streamindex))
+void ui3BehaviorBase::OnStreamSample(svlSample * CMN_UNUSED(sample), int CMN_UNUSED(streamindex))
 {
     // Default implementation does nothing
 }
 
 
-size_t ui3BehaviorBase::AddStream(svlStreamType type, const std::string & streamname)
+int ui3BehaviorBase::AddStream(svlStreamType type, const std::string & streamname)
 {
-    size_t streamindex = Streams.size();
+    int streamindex = Streams.size();
     ui3VideoInterfaceFilter* filter = new ui3VideoInterfaceFilter(type, streamindex, this);
     CMN_ASSERT(filter);
 
@@ -232,7 +232,7 @@ size_t ui3BehaviorBase::AddStream(svlStreamType type, const std::string & stream
 }
 
 
-unsigned int ui3BehaviorBase::GetStreamWidth(const size_t streamindex, unsigned int channel)
+unsigned int ui3BehaviorBase::GetStreamWidth(const int streamindex, unsigned int channel)
 {
     if (streamindex < 0 || streamindex >= static_cast<int>(Streams.size())) return 0;
     return Streams[streamindex]->GetWidth(channel);
